pyproject.toml
# pyproject.toml
[tool.poetry]
name = "Linkedin scraper"
version = "0.1.0"
description = "A job scraper for filtering job postings published on Linkedin"
authors = ["Osvaldo Colina <osvaldo.colina@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
pandas = "^2.2.0"
python-dotenv = "^1.0.0"
selenium = "^4.28.1"
sqlalchemy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
linkedin_scraper = "linkedin_scraper.scraper:main"
